✅ 21TB3J code: done
Task: 202604210859-21TB3J
Primary: code
Agent: CODER
Author: CODER
Status: DONE
Comment: verified: resolved recipe manifest/types hotspot as hand-maintained code with a domain split, kept public types barrel stable, and passed recipes typecheck/tests/build plus types-file guard.
