+++ packages/recipes/package.json
@@ -48,6 +48,6 @@
     "typecheck": "tsc -b",
     "prepack": "bun run clean && bun run build",
     "prepublishOnly": "node ../../scripts/enforce-github-publish.mjs && npm run prepack",
-    "test": "node --test"
+    "test": "bun test ./src/index.test.ts ./src/overlay.test.ts"
   }
 }
