+++ packages/recipes/src/manifest.ts
@@ -1,11 +1,13 @@
 import { readFile } from "node:fs/promises";
 
+import { invalidFieldMessage, isRecord } from "./internal-utils.js";
 import {
-  dedupeStrings,
-  invalidFieldMessage,
-  isRecord,
-  requiredFieldMessage,
-} from "./internal-utils.js";
+  normalizeNumber,
+  normalizeOptionalString,
+  normalizeOptionalStringList,
+  normalizeRequiredString,
+  normalizeStringList,
+} from "./manifest-primitives.js";
 import {
   normalizeAgentId,
   normalizeRecipeId,
@@ -30,48 +32,6 @@ import type {
   RecipeSkillDefinition,
   RecipeToolDefinition,
 } from "./types.js";
-
-function normalizeRequiredString(raw: unknown, field: string): string {
-  if (typeof raw !== "string") throw new Error(invalidFieldMessage(field, "string"));
-  const value = raw.trim();
-  if (!value) throw new Error(requiredFieldMessage(field));
-  return value;
-}
-
-function normalizeOptionalString(raw: unknown, field: string): string | undefined {
-  if (raw === undefined) return undefined;
-  if (typeof raw !== "string") throw new Error(invalidFieldMessage(field, "string"));
-  const value = raw.trim();
-  return value || undefined;
-}
-
-function normalizeStringList(raw: unknown, field: string, opts?: { minLength?: number }): string[] {
-  if (!Array.isArray(raw)) throw new Error(invalidFieldMessage(field, "string[]"));
-  const values = raw.map((value) => {
-    if (typeof value !== "string") throw new Error(invalidFieldMessage(field, "string[]"));
-    const trimmed = value.trim();
-    if (!trimmed) throw new Error(invalidFieldMessage(field, "string[]"));
-    return trimmed;
-  });
-  const deduped = dedupeStrings(values);
-  if ((opts?.minLength ?? 0) > 0 && deduped.length < (opts?.minLength ?? 0)) {
-    throw new Error(invalidFieldMessage(field, `string[${opts?.minLength ?? 0}+]`));
-  }
-  return deduped;
-}
-
-function normalizeOptionalStringList(raw: unknown, field: string): string[] | undefined {
-  if (raw === undefined) return undefined;
-  return normalizeStringList(raw, field);
-}
-
-function normalizeNumber(raw: unknown, field: string): number | undefined {
-  if (raw === undefined) return undefined;
-  if (typeof raw !== "number" || Number.isNaN(raw)) {
-    throw new Error(invalidFieldMessage(field, "number"));
-  }
-  return raw;
-}
 
 function normalizeCompatibility(raw: unknown): RecipeCompatibility | undefined {
   if (raw === undefined) return undefined;
